refactor: e2e_support wave-3 Slice B lib topic splits under 300 lines
Split hit-testable scroll, scenario opener, naval move pick, aliases_ui,
and fleet-reach preamble into topic modules re-exported from stable paths
so Slice B targets clear the forthcoming 300-line lib ceiling (Refs #4344).
