+++ packages/colonizethis_app_e2e_support/lib/e2e_helpers_aliases_ui.dart
@@ -1,296 +1,5 @@
-/// AC1 facade aliases extracted from [e2e_helpers.dart] (Refs #4075 AC3).
+/// AC1 facade aliases extracted from [e2e_helpers.dart] (Refs #4075 AC3 / #4344 Slice B).
 library;
 
-import 'package:colonizethis_app_l10n/l10n/app_localizations_contract.dart';
-import 'package:flutter/widgets.dart';
-import 'package:flutter_test/flutter_test.dart';
-
-import 'package:colonizethis_app_e2e_support/e2e_test_shared.dart';
-import 'package:colonizethis_app_e2e_support/e2e_test_shared_bootstrap.dart';
-
-Future<void> pumpFor(WidgetTester tester, Duration total) =>
-    e2ePumpFor(tester, total);
-
-Future<void> waitUntilFound(
-  WidgetTester tester,
-  Finder finder, {
-  required Duration timeout,
-  Duration diagnoseAfter = Duration.zero,
-  E2ePerfLog? perf,
-  String phaseName = 'wait_until_found',
-}) => e2eWaitUntilFound(
-  tester,
-  finder,
-  timeout: timeout,
-  diagnoseAfter: diagnoseAfter,
-  perf: perf,
-  phaseName: phaseName,
-);
-
-Future<void> dismissTransientUi(WidgetTester tester, {E2ePerfLog? perf}) =>
-    e2eDismissTransientUi(tester, perf: perf);
-
-/// AC1 alias for [e2eEnsureVisibleAndTapHitTestable] (Refs #2336 / #4075).
-/// AC1 alias for [e2eEnsureVisibleAndTapHitTestable] (Refs #2336 / #4075).
-Future<bool> ensureVisibleAndTapHitTestable(
-  WidgetTester tester,
-  Finder trigger,
-) => e2eEnsureVisibleAndTapHitTestable(tester, trigger);
-
-/// AC1 alias for [e2eAwaitPanelMountAfterOpenerTap] (Refs #2336 / #4075).
-/// AC1 alias for [e2eAwaitPanelMountAfterOpenerTap] (Refs #2336 / #4075).
-Future<bool> awaitPanelMountAfterOpenerTap(
-  WidgetTester tester,
-  Finder panelRoot, {
-  required Duration timeout,
-  E2ePerfLog? perf,
-  required String phaseName,
-}) => e2eAwaitPanelMountAfterOpenerTap(
-  tester,
-  panelRoot,
-  timeout: timeout,
-  perf: perf,
-  phaseName: phaseName,
-);
-
-/// AC1 alias for [e2eClosePanelOpenerSheetAndAwaitOpener] (Refs #2336 / #4075).
-/// AC1 alias for [e2eClosePanelOpenerSheetAndAwaitOpener] (Refs #2336 / #4075).
-Future<void> closePanelOpenerSheetAndAwaitOpener(
-  WidgetTester tester, {
-  required Finder primary,
-  Finder? secondary,
-  required String afterSheetClearPhase,
-  required String awaitOpenerPhase,
-  E2ePerfLog? perf,
-  Duration bottomSheetCloseTimeout = kE2eDefaultBottomSheetCloseTimeout,
-  Duration sheetClearTimeout = const Duration(seconds: 2),
-  Duration awaitOpenerTimeout = const Duration(seconds: 3),
-}) => e2eClosePanelOpenerSheetAndAwaitOpener(
-  tester,
-  primary: primary,
-  secondary: secondary,
-  afterSheetClearPhase: afterSheetClearPhase,
-  awaitOpenerPhase: awaitOpenerPhase,
-  perf: perf,
-  bottomSheetCloseTimeout: bottomSheetCloseTimeout,
-  sheetClearTimeout: sheetClearTimeout,
-  awaitOpenerTimeout: awaitOpenerTimeout,
-);
-
-/// AC1 alias for [e2eOpenerTapTriggerAndAwaitMount] (Refs #2336 / #4075).
-/// AC1 alias for [e2eOpenerTapTriggerAndAwaitMount] (Refs #2336 / #4075).
-Future<bool> openerTapTriggerAndAwaitMount(
-  WidgetTester tester, {
-  required Finder trigger,
-  required Finder panelRoot,
-  required Duration mountTimeout,
-  required String mountPhaseName,
-  E2ePerfLog? perf,
-}) => e2eOpenerTapTriggerAndAwaitMount(
-  tester,
-  trigger: trigger,
-  panelRoot: panelRoot,
-  mountTimeout: mountTimeout,
-  mountPhaseName: mountPhaseName,
-  perf: perf,
-);
-
-/// AC1 alias for [e2eOpenPanelViaRailOrMarker] (Refs #2336 / #4075).
-/// AC1 alias for [e2eOpenPanelViaRailOrMarker] (Refs #2336 / #4075).
-Future<void> openPanelViaRailOrMarker(
-  WidgetTester tester, {
-  required String openerLabel,
-  required Finder railButton,
-  required Finder markerButton,
-  required Finder panelRoot,
-  required String afterSheetPanelsClearPhase,
-  required String Function(Duration timeout) timeoutMessageBuilder,
-  Duration overallTimeout = const Duration(seconds: 20),
-  Duration bottomSheetCloseTimeout = kE2eDefaultBottomSheetCloseTimeout,
-  Duration mountTimeout = const Duration(seconds: 3),
-  E2ePerfLog? perf,
-}) => e2eOpenPanelViaRailOrMarker(
-  tester,
-  openerLabel: openerLabel,
-  railButton: railButton,
-  markerButton: markerButton,
-  panelRoot: panelRoot,
-  afterSheetPanelsClearPhase: afterSheetPanelsClearPhase,
-  timeoutMessageBuilder: timeoutMessageBuilder,
-  overallTimeout: overallTimeout,
-  bottomSheetCloseTimeout: bottomSheetCloseTimeout,
-  mountTimeout: mountTimeout,
-  perf: perf,
-);
-
-/// AC1 alias for [e2eDismissAlertDialogIfPresent] (Refs #2336 / #4075).
-/// AC1 alias for [e2eDismissAlertDialogIfPresent] (Refs #2336 / #4075).
-Future<bool> dismissAlertDialogIfPresent(
-  WidgetTester tester, {
-  E2ePerfLog? perf,
-  Duration dismissTimeout = kE2eDefaultAlertDialogDismissTimeout,
-  List<String> dismissLabels = kE2eDefaultAlertDialogDismissLabels,
-}) => e2eDismissAlertDialogIfPresent(
-  tester,
-  perf: perf,
-  dismissTimeout: dismissTimeout,
-  dismissLabels: dismissLabels,
-);
-
-/// AC1 alias for [e2eDismissGenericOkIfPresent] (Refs #2336 / #4075).
-/// AC1 alias for [e2eDismissGenericOkIfPresent] (Refs #2336 / #4075).
-Future<bool> dismissGenericOkIfPresent(
-  WidgetTester tester, {
-  E2ePerfLog? perf,
-  Duration dismissTimeout = kE2eDefaultGenericOkDismissTimeout,
-  String label = kE2eDefaultGenericOkLabel,
-}) => e2eDismissGenericOkIfPresent(
-  tester,
-  perf: perf,
-  dismissTimeout: dismissTimeout,
-  label: label,
-);
-
-/// AC1 alias for [e2eDismissCtDialogShellBroadSweepIfPresent] (Refs #2336 / #4075).
-/// AC1 alias for [e2eDismissCtDialogShellBroadSweepIfPresent] (Refs #2336 / #4075).
-Future<bool> dismissCtDialogShellBroadSweepIfPresent(
-  WidgetTester tester, {
-  E2ePerfLog? perf,
-  Duration dismissTimeout = kE2eDefaultCtDialogShellBroadSweepDismissTimeout,
-}) => e2eDismissCtDialogShellBroadSweepIfPresent(
-  tester,
-  perf: perf,
-  dismissTimeout: dismissTimeout,
-);
-
-/// AC1 alias for [e2eDismissCtDialogShellIfPresent] (Refs #2336 / #4075).
-/// AC1 alias for [e2eDismissCtDialogShellIfPresent] (Refs #2336 / #4075).
-Future<bool> dismissCtDialogShellIfPresent(
-  WidgetTester tester,
-  AppLocalizations l10n, {
-  E2ePerfLog? perf,
-  Duration shellCloseTimeout = kE2eDefaultCtDialogShellCloseTimeout,
-  String phaseName = kE2eDefaultCtDialogShellClosePhase,
-}) => e2eDismissCtDialogShellIfPresent(
-  tester,
-  l10n,
-  perf: perf,
-  shellCloseTimeout: shellCloseTimeout,
-  phaseName: phaseName,
-);
-
-/// AC1 alias for [e2eDismissCtDialogShellWithPopRouteEscalation] (Refs #2336 / #4075).
-/// AC1 alias for [e2eDismissCtDialogShellWithPopRouteEscalation] (Refs #2336 / #4075).
-Future<bool> dismissCtDialogShellWithPopRouteEscalation(
-  WidgetTester tester, {
-  E2ePerfLog? perf,
-  Duration escalationTimeout = kE2eDefaultCtDialogShellEscalationTimeout,
-  String escalationPhase = kE2eDefaultCtDialogShellEscalationPhase,
-}) => e2eDismissCtDialogShellWithPopRouteEscalation(
-  tester,
-  perf: perf,
-  escalationTimeout: escalationTimeout,
-  escalationPhase: escalationPhase,
-);
-
-/// AC1 alias for [e2eDismissSnackBarIfPresent] (Refs #2336 / #4075).
-/// AC1 alias for [e2eDismissSnackBarIfPresent] (Refs #2336 / #4075).
-Future<bool> dismissSnackBarIfPresent(
-  WidgetTester tester, {
-  E2ePerfLog? perf,
-  Duration dismissTimeout = kE2eDefaultSnackBarDismissTimeout,
-}) => e2eDismissSnackBarIfPresent(
-  tester,
-  perf: perf,
-  dismissTimeout: dismissTimeout,
-);
-
-Future<void> openCivilianPanel(
-  WidgetTester tester, {
-  Duration timeout = const Duration(seconds: 20),
-  E2ePerfLog? perf,
-  Duration bottomSheetCloseTimeout = kE2eDefaultBottomSheetCloseTimeout,
-  String afterSheetPanelsClearPhase =
-      'pump_until_panels_cleared_after_close_sheet_civilian_open',
-}) => e2eOpenCivilianPanel(
-  tester,
-  timeout: timeout,
-  perf: perf,
-  bottomSheetCloseTimeout: bottomSheetCloseTimeout,
-  afterSheetPanelsClearPhase: afterSheetPanelsClearPhase,
-);
-
-Future<void> openNavalPanel(
-  WidgetTester tester, {
-  E2ePerfLog? perf,
-  Duration timeout = kE2eDefaultNavalOpenTimeout,
-  Duration bottomSheetCloseTimeout = kE2eDefaultBottomSheetCloseTimeout,
-}) => e2eOpenNavalPanel(
-  tester,
-  perf: perf,
-  timeout: timeout,
-  bottomSheetCloseTimeout: bottomSheetCloseTimeout,
-);
-
-Future<void> openProductionPanel(
-  WidgetTester tester, {
-  E2ePerfLog? perf,
-  Duration timeout = const Duration(seconds: 20),
-}) => e2eOpenProductionPanel(tester, perf: perf, timeout: timeout);
-
-Future<void> openPanelFromMarker(
-  WidgetTester tester, {
-  required Finder markerButton,
-  required Finder panelRoot,
-  Duration timeout = const Duration(seconds: 20),
-  E2ePerfLog? perf,
-}) => e2eOpenPanelFromMarker(
-  tester,
-  markerButton: markerButton,
-  panelRoot: panelRoot,
-  timeout: timeout,
-  perf: perf,
-);
-
-Future<Duration> waitForNextTurnLabelAdvance(
-  WidgetTester tester, {
-  required String turnLabelBefore,
-  required Duration timeout,
-  E2ePerfLog? perf,
-}) => e2eWaitForNextTurnLabelAdvance(
-  tester,
-  turnLabelBefore: turnLabelBefore,
-  timeout: timeout,
-  perf: perf,
-);
-
-Future<Duration> advanceOneHumanTurn(
-  WidgetTester tester, {
-  required AppLocalizations l10n,
-  E2ePerfLog? perf,
-  Duration timeout = kE2eNextTurnResolutionTimeout,
-}) => e2eAdvanceOneHumanTurn(tester, l10n: l10n, perf: perf, timeout: timeout);
-
-Future<void> closeBottomSheet(
-  WidgetTester tester, {
-  E2ePerfLog? perf,
-  Duration overallTimeout = kE2eDefaultBottomSheetCloseTimeout,
-}) => e2eCloseBottomSheet(tester, perf: perf, overallTimeout: overallTimeout);
-
-Future<void> bootstrapNewGameToMap(
-  WidgetTester tester, {
-  E2ePerfLog? perf,
-  Duration overallCap = const Duration(seconds: 60),
-  String? advancedStartOptionLabel,
-}) => e2eBootstrapNewGameToMap(
-  tester,
-  perf: perf,
-  overallCap: overallCap,
-  advancedStartOptionLabel: advancedStartOptionLabel,
-);
-
-void collectTextPreorder(Element element, List<String> out) =>
-    e2eCollectTextPreorder(element, out);
-
-Future<void> expandEachExpansionTileOnce(WidgetTester tester) =>
-    e2eExpandEachExpansionTileOnce(tester);
+export 'e2e_helpers_aliases_ui_panels.dart';
+export 'e2e_helpers_aliases_ui_wait.dart';
